Move getRoofwindowOpenLevel to itemutils
