+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
     "standard.enableGlobally": true,
-    "standard.engine": "semistandard",
+    "standard.engine": "standardx",
     "standard.trace.server": "messages",
     "markdown.extension.toc.levels": "2..4",
     "git.alwaysSignOff": true
