+++ src/index.js
@@ -1,5 +1,4 @@
 module.exports = {
   get rulesx () { return require('./rules'); },
-  get itemutils () { return require('./itemutils'); },
-  get utils () { return require('./utils'); }
+  get itemutils () { return require('./itemutils'); }
 };
