+++ src/itemutils/windows.js
@@ -11,29 +11,21 @@
 const { items } = require('openhab');
 
 /**
- * Utils namespace
- *
- * This namespace provides utilities for use in this package and your own rules.
- * Some of them are specific to my use case!
- * @namespace utils
- */
+  * @typedef {Object} roofwindowState representation of roofwindow state / openess level
+  * @memberof itemutils
+  * @property {String} text textual representation of openess level (in German)
+  * @property {Number} int Integer reprentation of openess level: 100 for completely closed, 1 for a little bit opened, 2 for a bit opened & 4 for completely opened (5 on error)
+  */
 
 /**
- * @typedef {Object} roofwindowState representation of roofwindow state / openess level
- * @memberof utils
- * @property {String} text textual representation of openess level (in German)
- * @property {Number} int Integer reprentation of openess level: 100 for completely closed, 1 for a little bit opened, 2 for a bit opened & 4 for completely opened (5 on error)
- */
-
-/**
- * Get the level a roofwindow is opened.
- * The roofwindow can either be completely opened or a bit opened (große Lüftung) or a little bit opened (kleine Lüftung) or closed.
- * Item naming scheme is required: `baseItem` + `_zu`, `_klLueftung`, `grLueftung`.
- *
- * @memberof utils
- * @param {String} baseItem base of the Items names, e.g. `Florian_Dachfenster`
- * @returns {utils.roofwindowState}
- */
+  * Get the level a roofwindow is opened.
+  * The roofwindow can either be completely opened or a bit opened (große Lüftung) or a little bit opened (kleine Lüftung) or closed.
+  * Item naming scheme is required: `baseItem` + `_zu`, `_klLueftung`, `grLueftung`.
+  *
+  * @memberof itemutils
+  * @param {String} baseItem base of the Items names, e.g. `Florian_Dachfenster`
+  * @returns {itemutils.roofwindowState}
+  */
 const getRoofwindowOpenLevel = (baseItem) => {
   const output = {};
   const stateClosed = items.getItem(baseItem + '_zu').state;
